backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
email-validator
python-multipart
bcrypt
python-jose[cryptography]
httpx
